+++ lib/elixir/test/elixir/regex_test.exs
@@ -149,14 +149,16 @@ defmodule RegexTest do
 
   test :split do
     assert Regex.split(~r",", "") == [""]
+    assert Regex.split(~r",", "", trim: true) == [""]
     assert Regex.split(~r" ", "foo bar baz") == ["foo", "bar", "baz"]
-    assert Regex.split(~r" ", "foo bar baz", parts: 0) == ["foo", "bar", "baz"]
     assert Regex.split(~r" ", "foo bar baz", parts: :infinity) == ["foo", "bar", "baz"]
     assert Regex.split(~r" ", "foo bar baz", parts: 10) == ["foo", "bar", "baz"]
     assert Regex.split(~r" ", "foo bar baz", parts: 2) == ["foo", "bar baz"]
     assert Regex.split(~r"\s", "foobar") == ["foobar"]
     assert Regex.split(~r" ", " foo bar baz ") == ["", "foo", "bar", "baz", ""]
     assert Regex.split(~r" ", " foo bar baz ", trim: true) == ["foo", "bar", "baz"]
+    assert Regex.split(~r" ", " foo bar baz ", parts: 2) == ["", "foo bar baz "]
+    assert Regex.split(~r" ", " foo bar baz ", trim: true, parts: 2) == ["foo", "bar baz "]
     assert Regex.split(~r"=", "key=") == ["key", ""]
     assert Regex.split(~r"=", "=value") == ["", "value"]
   end